backend/requirements.txt
Django>=4.2,<5.0
PyMySQL>=1.1.0
django-cors-headers>=4.0.0
PyJWT>=2.8.0
requests>=2.31.0
python-dotenv>=1.0.0
